created data.sql to prepopulate demo data and configured app to recreate schema on startup
Instead of saving data inbetween app startup/closes, the database now gets refreshed and populated with this set of demo data. In the demo data, club with id 1 is the most populated, and user with id 3 is the member with the most booking history. Also includes a gym class (id 14) at max capacity, and more.